refactor(types-registry): consolidate test store decorators
- Share forwarding implementations through TestStores hooks.
- Verify backend lock waits during revision race tests.
- Update test documentation and call sites.
