+++ gears/system/types-registry/types-registry/tests/common/test_stores.rs
@@ -1,0 +1,567 @@
+//! Test hooks over the real persistence adapter, with shared port forwarding.
+//!
+//! [`TestStores<H>`] uses [`PauseHooks`] to pause a call, [`ClaimHooks`] to signal
+//! claim entry and return, or [`CasMissHooks`] to refuse a compare-and-swap.
+//!
+//! Interception points are shared: add a [`PausePoint`] and call site for timing,
+//! or a [`StoreHooks`] method to inspect arguments/results or override a result.
+//!
+//! `async_trait` boxes each executed `at` call, including no-op hooks, unless
+//! optimized away. The no-op does not yield; allocation overhead is unmeasured.
+
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use time::OffsetDateTime;
+use toolkit_db::DbTx;
+use toolkit_db::secure::{AccessScope, ScopeError};
+use types_registry::domain::admission::fingerprint::ScopeHash;
+use types_registry::domain::enums::{DependencyKind, EntityKind, OwnershipScope};
+use types_registry::domain::family::FamilyKey;
+use types_registry::domain::ports::{
+    CurrentDocument, CurrentInstanceRow, CurrentInstanceValue, CurrentSchemaCas,
+    CurrentSchemaProjection, CurrentTypeSchemaRow, DependencyClosure, DependencyStore, EntityRow,
+    EntityStore, EntityWriteOrderStore, InstanceStore, NewCurrentInstance, NewCurrentTypeSchema,
+    NewEntity, NewInstanceRevision, NewOperation, NewOperationItem, NewRevision, OperationItemRow,
+    OperationRow, OperationStore, ReverseImpact, Stores, TypeSchemaStore, VersionFamilyRow,
+    VersionFamilyStore,
+};
+use uuid::Uuid;
+
+use super::stores;
+
+/// Hook locations in the commit transaction.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum PausePoint {
+    /// Before the commit's first statement claims `entity_write_order`.
+    BeforeEntityWriteOrderClaim,
+    /// After the claim succeeds.
+    AfterEntityWriteOrderClaim,
+    /// After the entity/content reads, before the unchanged re-read or CAS.
+    CurrentDocuments,
+    /// After creation takes the family row but before checking its rules.
+    CreateOrGet,
+    RevisionEntityRead,
+}
+
+/// Port-call hooks with no-op defaults.
+#[async_trait]
+pub trait StoreHooks: Send + Sync {
+    /// Runs at each reached [`PausePoint`]; may hold the transaction open.
+    async fn at(&self, _point: PausePoint) {}
+
+    /// Return `true` to simulate a schema CAS miss without a database write.
+    fn refuse_schema_cas(&self, _entity_id: i64) -> bool {
+        false
+    }
+}
+
+/// Pauses one matching call until the test resumes it.
+pub struct PauseHooks {
+    at: PausePoint,
+    /// Matching call to pause, starting at 1.
+    nth: usize,
+    seen: std::sync::atomic::AtomicUsize,
+    reached: tokio::sync::Mutex<Option<tokio::sync::oneshot::Sender<()>>>,
+    resume: tokio::sync::Mutex<Option<tokio::sync::oneshot::Receiver<()>>>,
+}
+
+#[async_trait]
+impl StoreHooks for PauseHooks {
+    /// Signal and pause only the selected occurrence.
+    async fn at(&self, point: PausePoint) {
+        if point != self.at {
+            return;
+        }
+        if self.seen.fetch_add(1, std::sync::atomic::Ordering::SeqCst) + 1 != self.nth {
+            return;
+        }
+        let reached = self.reached.lock().await.take();
+        let resume = self.resume.lock().await.take();
+        if let Some(reached) = reached {
+            // Ignore a receiver dropped by an aborted test.
+            reached.send(()).ok();
+        }
+        if let Some(resume) = resume {
+            resume.await.expect("the test must always resume the pass");
+        }
+    }
+}
+
+/// Signals claim entry and successful return. A lock wait must be verified
+/// separately; see `assert_backend_reports_a_blocked_claim` in the backend tests.
+pub struct ClaimHooks {
+    entered: tokio::sync::Mutex<Option<tokio::sync::oneshot::Sender<()>>>,
+    returned: tokio::sync::Mutex<Option<tokio::sync::oneshot::Sender<()>>>,
+}
+
+#[async_trait]
+impl StoreHooks for ClaimHooks {
+    async fn at(&self, point: PausePoint) {
+        let slot = match point {
+            PausePoint::BeforeEntityWriteOrderClaim => &self.entered,
+            PausePoint::AfterEntityWriteOrderClaim => &self.returned,
+            _ => return,
+        };
+        if let Some(signal) = slot.lock().await.take() {
+            // Ignore a receiver dropped by an aborted test.
+            signal.send(()).ok();
+        }
+    }
+}
+
+/// Refuses every schema CAS for one entity, including retries, to test rollback.
+pub struct CasMissHooks {
+    refuse_for_entity_id: i64,
+}
+
+impl StoreHooks for CasMissHooks {
+    fn refuse_schema_cas(&self, entity_id: i64) -> bool {
+        entity_id == self.refuse_for_entity_id
+    }
+}
+
+/// Real persistence adapter decorated with `H`'s hooks.
+pub struct TestStores<H> {
+    inner: Arc<dyn Stores>,
+    hooks: H,
+}
+
+impl TestStores<PauseHooks> {
+    /// Returns decorated ports, a pause notification, and a resume sender.
+    #[must_use]
+    pub fn pausing(
+        at: PausePoint,
+    ) -> (
+        Arc<Self>,
+        tokio::sync::oneshot::Receiver<()>,
+        tokio::sync::oneshot::Sender<()>,
+    ) {
+        Self::pausing_at_occurrence(at, 1)
+    }
+
+    /// Like [`Self::pausing`], but hold the `nth` matching call.
+    #[must_use]
+    pub fn pausing_at_occurrence(
+        at: PausePoint,
+        nth: usize,
+    ) -> (
+        Arc<Self>,
+        tokio::sync::oneshot::Receiver<()>,
+        tokio::sync::oneshot::Sender<()>,
+    ) {
+        let (reached_tx, reached_rx) = tokio::sync::oneshot::channel();
+        let (resume_tx, resume_rx) = tokio::sync::oneshot::channel();
+        let decorated = Arc::new(Self {
+            inner: stores(),
+            hooks: PauseHooks {
+                at,
+                nth,
+                seen: std::sync::atomic::AtomicUsize::new(0),
+                reached: tokio::sync::Mutex::new(Some(reached_tx)),
+                resume: tokio::sync::Mutex::new(Some(resume_rx)),
+            },
+        });
+        (decorated, reached_rx, resume_tx)
+    }
+}
+
+impl TestStores<ClaimHooks> {
+    /// Returns decorated ports and notifications for claim entry and success.
+    #[must_use]
+    pub fn claim_signalling() -> (
+        Arc<Self>,
+        tokio::sync::oneshot::Receiver<()>,
+        tokio::sync::oneshot::Receiver<()>,
+    ) {
+        let (entered_tx, entered_rx) = tokio::sync::oneshot::channel();
+        let (returned_tx, returned_rx) = tokio::sync::oneshot::channel();
+        (
+            Arc::new(Self {
+                inner: stores(),
+                hooks: ClaimHooks {
+                    entered: tokio::sync::Mutex::new(Some(entered_tx)),
+                    returned: tokio::sync::Mutex::new(Some(returned_tx)),
+                },
+            }),
+            entered_rx,
+            returned_rx,
+        )
+    }
+}
+
+impl TestStores<CasMissHooks> {
+    /// Refuse `refuse_for_entity_id`'s current-schema compare-and-swap.
+    #[must_use]
+    pub fn cas_miss(refuse_for_entity_id: i64) -> Arc<Self> {
+        Arc::new(Self {
+            inner: stores(),
+            hooks: CasMissHooks {
+                refuse_for_entity_id,
+            },
+        })
+    }
+}
+
+// Port implementations.
+
+#[async_trait]
+impl<H: StoreHooks> EntityWriteOrderStore for TestStores<H> {
+    async fn claim_entity_write_order(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        now: OffsetDateTime,
+    ) -> Result<(), ScopeError> {
+        self.hooks.at(PausePoint::BeforeEntityWriteOrderClaim).await;
+        self.inner.claim_entity_write_order(tx, scope, now).await?;
+        self.hooks.at(PausePoint::AfterEntityWriteOrderClaim).await;
+        Ok(())
+    }
+}
+
+#[async_trait]
+impl<H: StoreHooks> VersionFamilyStore for TestStores<H> {
+    async fn create_or_get(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        family_key: &FamilyKey,
+        ownership_scope: OwnershipScope,
+        owner_tenant_id: Option<Uuid>,
+        now: OffsetDateTime,
+    ) -> Result<(VersionFamilyRow, bool), ScopeError> {
+        let out = self
+            .inner
+            .create_or_get(tx, scope, family_key, ownership_scope, owner_tenant_id, now)
+            .await?;
+        self.hooks.at(PausePoint::CreateOrGet).await;
+        Ok(out)
+    }
+}
+
+#[async_trait]
+impl<H: StoreHooks> EntityStore for TestStores<H> {
+    async fn find_by_gts_id(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        gts_id: &str,
+    ) -> Result<Option<EntityRow>, ScopeError> {
+        self.hooks.at(PausePoint::RevisionEntityRead).await;
+        self.inner.find_by_gts_id(tx, scope, gts_id).await
+    }
+
+    async fn find_by_gts_ids(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        gts_ids: &[String],
+    ) -> Result<Vec<EntityRow>, ScopeError> {
+        self.inner.find_by_gts_ids(tx, scope, gts_ids).await
+    }
+
+    async fn find_by_gts_uuid(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        gts_uuid: Uuid,
+    ) -> Result<Option<EntityRow>, ScopeError> {
+        self.inner.find_by_gts_uuid(tx, scope, gts_uuid).await
+    }
+
+    async fn kind_in_family(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        family_id: i64,
+    ) -> Result<Option<EntityKind>, ScopeError> {
+        self.inner.kind_in_family(tx, scope, family_id).await
+    }
+
+    async fn insert_entity(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewEntity,
+    ) -> Result<Option<EntityRow>, ScopeError> {
+        self.inner.insert_entity(tx, scope, new).await
+    }
+
+    async fn compare_and_swap_version(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        entity_id: i64,
+        expected_resource_version: i64,
+        now: OffsetDateTime,
+    ) -> Result<Option<i64>, ScopeError> {
+        self.inner
+            .compare_and_swap_version(tx, scope, entity_id, expected_resource_version, now)
+            .await
+    }
+}
+
+#[async_trait]
+impl<H: StoreHooks> TypeSchemaStore for TestStores<H> {
+    async fn current_documents(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        entity_ids: &[i64],
+    ) -> Result<Vec<CurrentDocument>, ScopeError> {
+        let out = self.inner.current_documents(tx, scope, entity_ids).await?;
+        self.hooks.at(PausePoint::CurrentDocuments).await;
+        Ok(out)
+    }
+
+    async fn find_current_schema(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        entity_id: i64,
+    ) -> Result<Option<CurrentTypeSchemaRow>, ScopeError> {
+        self.inner.find_current_schema(tx, scope, entity_id).await
+    }
+
+    async fn current_schema_projections(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        entity_ids: &[i64],
+    ) -> Result<Vec<CurrentSchemaProjection>, ScopeError> {
+        self.inner
+            .current_schema_projections(tx, scope, entity_ids)
+            .await
+    }
+
+    async fn insert_schema_revision(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewRevision,
+    ) -> Result<(), ScopeError> {
+        self.inner.insert_schema_revision(tx, scope, new).await
+    }
+
+    async fn insert_current_schema(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewCurrentTypeSchema,
+    ) -> Result<(), ScopeError> {
+        self.inner.insert_current_schema(tx, scope, new).await
+    }
+
+    async fn update_current_schema(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewCurrentTypeSchema,
+        expected: CurrentSchemaCas,
+    ) -> Result<bool, ScopeError> {
+        if self.hooks.refuse_schema_cas(new.entity_id) {
+            // Simulate the projection moving after its token was captured.
+            return Ok(false);
+        }
+        self.inner
+            .update_current_schema(tx, scope, new, expected)
+            .await
+    }
+}
+
+#[async_trait]
+impl<H: StoreHooks> InstanceStore for TestStores<H> {
+    async fn current_values(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        entity_ids: &[i64],
+    ) -> Result<Vec<CurrentInstanceValue>, ScopeError> {
+        self.inner.current_values(tx, scope, entity_ids).await
+    }
+
+    async fn find_current_instance(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        entity_id: i64,
+    ) -> Result<Option<CurrentInstanceRow>, ScopeError> {
+        self.inner.find_current_instance(tx, scope, entity_id).await
+    }
+
+    async fn insert_instance_revision(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewInstanceRevision,
+    ) -> Result<(), ScopeError> {
+        self.inner.insert_instance_revision(tx, scope, new).await
+    }
+
+    async fn insert_current_instance(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewCurrentInstance,
+    ) -> Result<(), ScopeError> {
+        self.inner.insert_current_instance(tx, scope, new).await
+    }
+
+    async fn update_current_instance(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewCurrentInstance,
+    ) -> Result<bool, ScopeError> {
+        self.inner.update_current_instance(tx, scope, new).await
+    }
+}
+
+#[async_trait]
+impl<H: StoreHooks> OperationStore for TestStores<H> {
+    async fn find_by_idempotency(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        idempotency_scope_hash: &ScopeHash,
+        idempotency_key: &str,
+    ) -> Result<Option<OperationRow>, ScopeError> {
+        self.inner
+            .find_by_idempotency(tx, scope, idempotency_scope_hash, idempotency_key)
+            .await
+    }
+
+    async fn find_by_id(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        id: Uuid,
+    ) -> Result<Option<OperationRow>, ScopeError> {
+        self.inner.find_by_id(tx, scope, id).await
+    }
+
+    async fn insert_operation(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        new: NewOperation,
+    ) -> Result<OperationRow, ScopeError> {
+        self.inner.insert_operation(tx, scope, new).await
+    }
+
+    async fn insert_items(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        parent: &OperationRow,
+        items: &[NewOperationItem],
+    ) -> Result<(), ScopeError> {
+        self.inner.insert_items(tx, scope, parent, items).await
+    }
+
+    async fn find_items(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        operation_id: Uuid,
+    ) -> Result<Vec<OperationItemRow>, ScopeError> {
+        self.inner.find_items(tx, scope, operation_id).await
+    }
+
+    async fn mark_running(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        id: Uuid,
+        now: OffsetDateTime,
+    ) -> Result<bool, ScopeError> {
+        self.inner.mark_running(tx, scope, id, now).await
+    }
+
+    async fn mark_completed(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        id: Uuid,
+        now: OffsetDateTime,
+    ) -> Result<bool, ScopeError> {
+        self.inner.mark_completed(tx, scope, id, now).await
+    }
+
+    async fn mark_item_succeeded(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        item_id: i64,
+        revision_no: i32,
+        resource_version: i64,
+        now: OffsetDateTime,
+    ) -> Result<bool, ScopeError> {
+        self.inner
+            .mark_item_succeeded(tx, scope, item_id, revision_no, resource_version, now)
+            .await
+    }
+
+    async fn mark_item_unchanged(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        item_id: i64,
+        resource_version: i64,
+        now: OffsetDateTime,
+    ) -> Result<bool, ScopeError> {
+        self.inner
+            .mark_item_unchanged(tx, scope, item_id, resource_version, now)
+            .await
+    }
+
+    async fn mark_item_failed(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        item_id: i64,
+        error_payload: String,
+        now: OffsetDateTime,
+    ) -> Result<bool, ScopeError> {
+        self.inner
+            .mark_item_failed(tx, scope, item_id, error_payload, now)
+            .await
+    }
+}
+
+#[async_trait]
+impl<H: StoreHooks> DependencyStore for TestStores<H> {
+    async fn closure(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        roots: &[String],
+    ) -> Result<DependencyClosure, ScopeError> {
+        self.inner.closure(tx, scope, roots).await
+    }
+
+    async fn reverse_impact(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        roots: &[i64],
+        write_set_bound: usize,
+    ) -> Result<ReverseImpact, ScopeError> {
+        self.inner
+            .reverse_impact(tx, scope, roots, write_set_bound)
+            .await
+    }
+
+    async fn replace_outgoing(
+        &self,
+        tx: &DbTx<'_>,
+        scope: &AccessScope,
+        from_entity_id: i64,
+        edges: &[(DependencyKind, i64)],
+    ) -> Result<(), ScopeError> {
+        self.inner
+            .replace_outgoing(tx, scope, from_entity_id, edges)
+            .await
+    }
+}
